feat: mask the closing active tab with a snapshot to hide swap flash
